[Refactor] : change hooks folder position
Take out the hooks folder from the utils folder.
